# Clean-up TinyProfiler output