--- a/src/Hipace.H
+++ b/src/Hipace.H
@@ -123,9 +123,6 @@ public:
      */
     void ResetAllQuantities ();
 
-    /** \brief reset all laser slices to 0 */
-    void ResetLaser ();
-
     /**
      * \brief does Coulomb collisions between plasmas and beams
      */