--- a/src/fields/fft_poisson_solver/MGPoissonSolverDirichlet.cpp
+++ b/src/fields/fft_poisson_solver/MGPoissonSolverDirichlet.cpp
@@ -17,8 +17,6 @@ MGPoissonSolverDirichlet::MGPoissonSolverDirichlet (
     amrex::DistributionMapping const& dm,
     amrex::Geometry const& gm )
 {
-    HIPACE_PROFILE("MGPoissonSolverDirichlet::define()");
-
     // need extra ghost cell for 2^n-1 HPMG
     m_stagingArea = amrex::MultiFab(ba, dm, 1, Fields::m_poisson_nguards + amrex::IntVect{1, 1, 0});
     AMREX_ALWAYS_ASSERT_WITH_MESSAGE(ba.size() == 1, "Parallel MG not supported");