--- a/src/fields/fft_poisson_solver/fft/WrapRocDST.cpp
+++ b/src/fields/fft_poisson_solver/fft/WrapRocDST.cpp
@@ -19,7 +19,6 @@ namespace AnyDST
     // see WrapCuDST for documentation
     void ExpandR2R (amrex::FArrayBox& dst, amrex::FArrayBox& src)
     {
-        HIPACE_DETAIL_PROFILE("AnyDST::ExpandR2R()");
         constexpr int scomp = 0;
         constexpr int dcomp = 0;
 
@@ -49,7 +48,6 @@ namespace AnyDST
     // see WrapCuDST for documentation
     void ShrinkC2R (amrex::FArrayBox& dst, amrex::BaseFab<amrex::GpuComplex<amrex::Real>>& src)
     {
-        HIPACE_DETAIL_PROFILE("AnyDST::ShrinkC2R()");
         constexpr int scomp = 0;
         constexpr int dcomp = 0;
 
@@ -71,7 +69,6 @@ namespace AnyDST
                     amrex::GpuComplex<amrex::Real>* const AMREX_RESTRICT out,
                     const int n_data, const int n_batch)
     {
-        HIPACE_DETAIL_PROFILE("AnyDST::ToComplex()");
         const int n_half = (n_data+1)/2;
         if((n_data%2 == 1)) {
             amrex::ParallelFor({{0,0,0}, {n_half,n_batch-1,0}},
@@ -105,7 +102,6 @@ namespace AnyDST
     void C2Rfft (AnyFFT::VendorFFTPlan& plan, amrex::GpuComplex<amrex::Real>* AMREX_RESTRICT in,
                  amrex::Real* const AMREX_RESTRICT out, rocfft_execution_info execinfo)
     {
-        HIPACE_DETAIL_PROFILE("AnyDST::C2Rfft()");
         rocfft_status result;
 
         void* in_arr[2] = {in, nullptr};
@@ -119,7 +115,6 @@ namespace AnyDST
     void ToSine (const amrex::Real* const AMREX_RESTRICT in, amrex::Real* const AMREX_RESTRICT out,
                  const int n_data, const int n_batch)
     {
-        HIPACE_DETAIL_PROFILE("AnyDST::ToSine()");
         using namespace amrex::literals;
 
         const amrex::Real n_1_real = n_data+1._rt;
@@ -149,7 +144,6 @@ namespace AnyDST
                     amrex::Real* const AMREX_RESTRICT out,
                     const int n_data, const int n_batch)
     {
-        HIPACE_DETAIL_PROFILE("AnyDST::Transpose()");
         constexpr int tile_dim = 32; //must be power of 2
         constexpr int block_rows = 8;
         const int num_blocks_x = (n_data + tile_dim - 1)/tile_dim;
@@ -367,8 +361,7 @@ namespace AnyDST
         rocfft_execution_info_destroy(dst_plan.m_execinfo);
     }
 
-    template<direction d>
-    void Execute (DSTplan& dst_plan){
+    void Execute (DSTplan& dst_plan, direction d){
         HIPACE_PROFILE("AnyDST::Execute()");
 
         if(!dst_plan.use_small_dst) {
@@ -425,7 +418,4 @@ namespace AnyDST
             Transpose(pos_arr, fourier_arr, ny, nx);
         }
     }
-
-    template void Execute<direction::forward>(DSTplan& dst_plan);
-    template void Execute<direction::backward>(DSTplan& dst_plan);
 }