--- a/src/utils/MultiBuffer.cpp
+++ b/src/utils/MultiBuffer.cpp
@@ -502,8 +502,6 @@ void MultiBuffer::put_data (int slice, MultiBeam& beams, MultiLaser& laser, int 
 }
 
 amrex::Real MultiBuffer::get_time () {
-    HIPACE_PROFILE("MultiBuffer::get_time()");
-
     if (m_is_serial) {
         return m_time_send_buffer;
     }
@@ -525,8 +523,6 @@ amrex::Real MultiBuffer::get_time () {
 }
 
 void MultiBuffer::put_time (amrex::Real time) {
-    HIPACE_PROFILE("MultiBuffer::put_time()");
-
     if (m_is_serial) {
         m_time_send_buffer = time;
         return;