@MaxThevenet 2024-04-08
Thanks for this PR! Is the cleaning specifically for GPU? Some functions may be negligible/predictable on GPU, but significant different on CPU.