@AlexanderSinn 2024-04-08
Good point. Looking at the profiling regions that were removed for CPU:
* `Hipace::ResetAllQuantities()` and `Hipace::ResetLaser()` are rarely called, omp parallelized and profiled by amrex through `FabArray::setVal()`.
* `Fields::LinCombination()` and `Fields::Multiply()` are omp parallelized and would now be profiled by the function that uses them e.g. `Fields::SolvePoissonPsiExmByEypBxEzBz()`.
* `FFTPoissonSolverPeriodic::define()` etc is replaced by `AnyFFT::CreatePlan()` since creating the FFT plan is the slow part on CPU when initializing the poisson solver.
* `AnyDST::ExpandR2R()` etc. is not used on CPU.
* `MultiPlasma::InitData()` is already profiled by the inner function `PlasmaParticleContainer::InitParticles()`
* `MultiBuffer::get_time()` and `MultiBuffer::put_time()` never take significant time on GPU runs, since they are doing MPI communication unrelated to GPU so I don’t see this being different for CPU simulations.